+++ mca-rust/common/Cargo.toml
@@ -10,3 +10,5 @@ crc = "3.0.0"
 nix = "0.25.0"
 libc = "0.2.132"
 more-asserts = "0.3.0"
+num_cpus = "1.13.1"
+systemstat = "0.2.0"
